[feature/parallel-requests] Refactoring: TrayImage umbenannt
