+++ src/test/java/com/github/funthomas424242/jenkinsmonitor/gui/JenkinsMonitorTrayTest.java
@@ -42,7 +42,7 @@ import java.io.IOException;
 import java.net.MalformedURLException;
 import java.net.URL;
 
-import static com.github.funthomas424242.jenkinsmonitor.gui.TrayImage.isImageOfColor;
+import static com.github.funthomas424242.jenkinsmonitor.gui.TrayImageTestHelper.isImageOfColor;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
