+++ src/test/java/com/github/funthomas424242/jenkinsmonitor/gui/TrayImageTestHelper.java
@@ -31,7 +31,7 @@ class ValueHolder {
     public boolean isOfColor = true;
 }
 
-public interface TrayImage {
+public interface TrayImageTestHelper {
 
     static boolean isImageOfColor(BufferedImage image, Color... colors) {
         final ValueHolder valueHolder = new ValueHolder();
